Use pip-compile to lock development dependencies
It's important to have a consistent development environment between
developers and CI so that dependency specific issues can be identified.

This is not the same as setting the versions that are used in production,
which are still bound by the constraints in setup.py

An alternative would be to use Poetry, which is probably a good idea anyway
to be honest as that would bring this project in line with other Kit Team
projects. (i.e good for volunteers as we have a more consistent dev experience)