Fix emulation of Zoloto spherical coordinates
This fixes a bug where the unpacking of the spherical coordinates
in Zoloto legacy mode was not actually compatible with Zoloto due
to the order of the tuple members having changed and the name of
one of the members had changed.

To preserve the new-style behaviour we break that out into its own
type with unchanged behaviour. The Zoloto emulation is replaced by
a copy of the logic which Zoloto itself used, slightly altered to
support the renamed member in both spellings.

The name of the (legacy) type is also changed here for consistency
with the other types names (it's now plural).

Fixes https://github.com/WillB97/april_vision/issues/8